+++ src/aioway/fn/ops/__init__.py
@@ -2,5 +2,6 @@
 
 "Torch operations, corresponding to `__torch_dispatch__` mode."
 
+from . import impls
 from .fn import *
 from .modes import *
